# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CountRequestsFullResponseDTO:
    properties:
      count:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
  dto.GetAnomaliesAmountDynamicsResponseDTO:
    properties:
      count:
        type: integer
      dynamics:
        items:
          $ref: '#/definitions/entity.AnomaliesDynamics'
        type: array
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dto.GetEfficiencyPlotResponseDTO:
    properties:
      error:
        type: string
      filename:
        type: string
      success:
        type: boolean
    type: object
...
  dto.GetRegionAreaResponseDTO:
    properties:
      area_polygon_geojson:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
...
  entity.AnomaliesDynamics:
    properties:
      amount:
        type: integer
      day:
        type: string
    type: object
...
host: ""
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: TechPotion ЛЦТ2022 Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger for TechPotion's leadershack2022 solution
  version: "1.0"
paths:
...
  /dashboard/anomalies/amount_dynamics:
    get:
      consumes:
      - application/json
      description: returns anomalies amount dynamics
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetAnomaliesAmountDynamicsResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.GetAnomaliesAmountDynamicsResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GetAnomaliesAmountDynamicsResponseDTO'
      summary: Get anomalies amount dynamics
      tags:
      - dashboard
...
  /dashboard/plots/efficiency:
    get:
      description: get an efficiency plot for dispatcher
      parameters:
      - description: Dispatcher (ОДС)
        in: query
        name: dispatcher
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetEfficiencyPlotResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GetEfficiencyPlotResponseDTO'
      summary: Get efficiency plot for dispatcher
      tags:
      - dashboard
...
  /regions/{region}/area:
    get:
      consumes:
      - application/json
      description: get region's area geojson polygon
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetRegionAreaResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.GetRegionAreaResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GetRegionAreaResponseDTO'
      summary: Get region area
      tags:
      - regions
...
  /requests_full/count:
    get:
      consumes:
      - application/json
      description: returns amount of requests full
      parameters:
      - description: Region
        in: query
        name: region
        required: true
        type: string
      - description: Lower time limit (RFC3339 formatted)
        in: query
        name: datetime_from
        required: true
        type: string
      - description: Upper time limit (RFC3339 formatted)
        in: query
        name: datetime_to
        required: true
        type: string
      - description: Bottom left X of screen
        in: query
        name: x_min
        type: number
      - description: Bottom left Y of screen
        in: query
        name: y_min
        type: number
      - description: Top right X of screen
        in: query
        name: x_max
        type: number
      - description: Top right Y of screen
        in: query
        name: y_max
        type: number
      - description: Urgency Category
        in: query
        name: urgency_category
        type: string
      - description: Anomaly Cases
        in: query
        items:
          type: integer
        name: anomaly_cases
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountRequestsFullResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CountRequestsFullResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.CountRequestsFullResponseDTO'
      summary: DEPRECATED Count requests fulls by selected area and region
      tags:
      - requests_full
...
schemes:
- http
swagger: "2.0"